feat(zCX): Add support for zCX from FP8
Make the changes needed so that s390x images can be built from experimental using FP8 (tested on both zLinux and zCX). No changes made to the ubi build as FP8 requires ubi8 and this doesn't currently work on zCX - in discussion with the zCX team and will submit a further PR once this is addressed.

Changed files:
   - experimental/ace-minimal-install/Dockerfile.ubuntu - remove arch dependency to allow to run on platforms other than x86

Contributes to Cloud-Integration/ace-build-and-test-pipeline/issues/334

Signed-off-by: Alison Lucas <alisonb@uk.ibm.com>